+++ examples/delayed.rs
@@ -1,4 +1,5 @@
 use std::process::exit;
+use std::thread::sleep;
 use std::time::Duration;
 use tyra::prelude::*;
 
@@ -6,11 +7,11 @@ use tyra::prelude::*;
 //Message//
 ///////////
 struct TemplateMessage {
-    id: usize
+    id: u32
 }
 
 impl TemplateMessage {
-    pub fn new(id: usize) -> Self {
+    pub fn new(id: u32) -> Self {
         Self {
             id
         }
@@ -54,7 +55,7 @@ impl ActorFactory<TemplateActor> for TemplateActorFactory {
 ///////////
 impl Handler<TemplateMessage> for TemplateActor {
     fn handle(&mut self, _msg: TemplateMessage, _context: &ActorContext<Self>) {
-        println!("SERS: {}", _msg.id)
+        println!("{}", _msg.id)
     }
 }
 
@@ -62,7 +63,6 @@ impl Handler<TemplateMessage> for TemplateActor {
 /////////
 //Usage//
 /////////
-
 fn main() {
     // first we need to get the config
     // this could be made mutable to add some dynamic config options
@@ -74,19 +74,14 @@ fn main() {
     let actor = actor_system.builder().spawn("template", TemplateActorFactory::new()).unwrap();
 
     // now we can send implemented messages to the actor
-    actor.send(TemplateMessage::new(1));
-    actor.send_after(TemplateMessage::new(2), Duration::from_secs(5));
-    actor.send_after(TemplateMessage::new(3), Duration::from_secs(4));
-    actor.send_after(TemplateMessage::new(4), Duration::from_secs(3));
-    actor.send_after(TemplateMessage::new(5), Duration::from_secs(2));
-    actor.send_after(TemplateMessage::new(6), Duration::from_secs(1));
-    actor.send(TemplateMessage::new(7));
-
-    //expected order:
-    //1, 7, 6, 5, 4, 3, 2
+    // we expect them to be handled in the order: 2, 3, 1
+    actor.send_after(TemplateMessage::new(1), Duration::from_secs(2));
+    actor.send(TemplateMessage::new(2));
+    actor.send_after(TemplateMessage::new(3), Duration::from_secs(1));
 
     // finally we ask the system to stop, which will in turn ask all actors to stop
     // exit code indicates if actor system was able to properly finish within the timeout
-    //actor_system.stop(Duration::from_secs(1));
+    sleep(Duration::from_secs(3));
+    actor_system.stop(Duration::from_secs(1));
     exit(actor_system.await_shutdown());
 }