+++ src/actor/actor_wrapper.rs
@@ -6,6 +6,8 @@ use crate::message::actor_stop_message::ActorStopMessage;
 use crate::prelude::Actor;
 use crate::system::wakeup_manager::WakeupManager;
 use std::panic::UnwindSafe;
+use std::time::Duration;
+use crate::system::internal_actor_manager::InternalActorManager;
 
 /// Wrapper used to interact with [Actor]
 pub struct ActorWrapper<A>
@@ -15,6 +17,7 @@ where
     mailbox: Mailbox<A>,
     address: ActorAddress,
     wakeup_manager: WakeupManager,
+    internal_actor_manager: Box<InternalActorManager>,
 }
 
 
@@ -25,11 +28,12 @@ where
     A: Actor + UnwindSafe,
 {
     /// Automatically called by the [ActorBuilder.build](../prelude/struct.ActorBuilder.html#method.build)
-    pub fn new(mailbox: Mailbox<A>, address: ActorAddress, wakeup_manager: WakeupManager) -> Self {
+    pub fn new(mailbox: Mailbox<A>, address: ActorAddress, wakeup_manager: WakeupManager, internal_actor_manager: InternalActorManager) -> Self {
         Self {
             mailbox,
             address,
             wakeup_manager,
+            internal_actor_manager: Box::new(internal_actor_manager)
         }
     }
 
@@ -47,6 +51,19 @@ where
         if self.mailbox.is_sleeping() {
             self.wakeup_manager.wakeup(self.address.clone());
         }
+    }
+
+    pub fn send_after<M>(&self, msg: M, delay: Duration)
+        where
+            A: Handler<M> + 'static,
+            M: ActorMessage + 'static,
+    {
+        if self.mailbox.is_stopped() {
+            return;
+        }
+
+        self.internal_actor_manager.send_after(msg, self.clone(), delay);
+
     }
 
     pub fn stop(&self) {
@@ -67,6 +84,7 @@ where
             wakeup_manager: self.wakeup_manager.clone(),
             mailbox: self.mailbox.clone(),
             address: self.address.clone(),
+            internal_actor_manager: self.internal_actor_manager.clone(),
         }
     }
 }