+++ src/routers/mod.rs
@@ -9,5 +9,6 @@ pub mod prelude {
     pub use crate::routers::bulk_router_message::BulkRouterMessage;
     pub use crate::routers::remove_actor_message::RemoveActorMessage;
     pub use crate::routers::round_robin_router::RoundRobinRouterFactory;
+    pub use crate::routers::round_robin_router::RoundRobinRouter;
     pub use crate::routers::router_message::RouterMessage;
 }